requirements.txt
PyQt5
PyAudio
python-dotenv
evdev
